Replace deprecated `StylesProvider` and `useStyles` with `createStylesContext`
- Consolidate files in order to make this change
- Remove `isFocused`/`data-focused` from `ClearIndicator`
- Remove all `ReactElement` return typings
- Unify the way the icons are defined
