+++ src/chakra-components/placeholder.tsx
@@ -1,5 +1,4 @@
 import React from "react";
-import type { ReactElement } from "react";
 import { Box } from "@chakra-ui/layout";
 import type { SystemStyleObject } from "@chakra-ui/system";
 import type { GroupBase, PlaceholderProps } from "react-select";
@@ -10,7 +9,7 @@ const Placeholder = <
   Group extends GroupBase<Option>
 >(
   props: PlaceholderProps<Option, IsMulti, Group>
-): ReactElement => {
+) => {
   const {
     children,
     className,
